chore: Convert GraphQL meeting types to use codegen
