+++ codegen.json
@@ -46,9 +46,12 @@
         "mappers": {
           "AcceptRequestToJoinDomainSuccess": "./types/AcceptRequestToJoinDomainSuccess#AcceptRequestToJoinDomainSuccessSource",
           "AcceptTeamInvitationPayload": "./types/AcceptTeamInvitationPayload#AcceptTeamInvitationPayloadSource",
+          "ActionMeeting": "../../database/types/MeetingAction#default",
+          "ActionMeetingMember": "../../database/types/ActionMeetingMember#default as ActionMeetingMemberDB",
           "AddApprovedOrganizationDomainsSuccess": "./types/AddApprovedOrganizationDomainsSuccess#AddApprovedOrganizationDomainsSuccessSource",
           "AddTranscriptionBotSuccess": "./types/AddTranscriptionBotSuccess#AddTranscriptionBotSuccessSource",
           "AddedNotification": "./types/AddedNotification#AddedNotificationSource",
+          "AgendaItem": "../../database/types/AgendaItem#default as AgendaItemDB",
           "ArchiveTeamPayload": "./types/ArchiveTeamPayload#ArchiveTeamPayloadSource",
           "AuthTokenPayload": "./types/AuthTokenPayload#AuthTokenPayloadSource",
           "AutogroupSuccess": "./types/AutogroupSuccess#AutogroupSuccessSource",
@@ -87,6 +90,7 @@
           "Organization": "../../database/types/Organization#default as Organization",
           "OrganizationUser": "../../database/types/OrganizationUser#default as OrganizationUser",
           "PokerMeeting": "../../database/types/MeetingPoker#default as MeetingPoker",
+          "PokerMeetingMember": "../../database/types/MeetingPokerMeetingMember#default as PokerMeetingMemberDB",
           "RRule": "rrule#RRule",
           "ReflectPrompt": "../../database/types/RetrospectivePrompt#default",
           "ReflectTemplate": "../../database/types/ReflectTemplate#default",
@@ -95,7 +99,9 @@
           "RemoveTeamMemberPayload": "./types/RemoveTeamMemberPayload#RemoveTeamMemberPayloadSource",
           "RequestToJoinDomainSuccess": "./types/RequestToJoinDomainSuccess#RequestToJoinDomainSuccessSource",
           "ResetReflectionGroupsSuccess": "./types/ResetReflectionGroupsSuccess#ResetReflectionGroupsSuccessSource",
+          "RetroReflectionGroup": "../../database/types/RetroReflectionGroup#default as RetroReflectionGroupDB",
           "RetrospectiveMeeting": "../../database/types/MeetingRetrospective#default",
+          "RetrospectiveMeetingMember": "../../database/types/RetroMeetingMember#default",
           "Reactable": "../../database/types/Reactable#Reactable",
           "RetrospectiveMeetingSettings": "../../database/types/MeetingSettingsRetrospective#default",
           "SAML": "./types/SAML#SAMLSource",
@@ -103,6 +109,8 @@
           "SetNotificationStatusPayload": "./types/SetNotificationStatusPayload#SetNotificationStatusPayloadSource",
           "SetOrgUserRoleSuccess": "./types/SetOrgUserRoleSuccess#SetOrgUserRoleSuccessSource",
           "ShareTopicSuccess": "./types/ShareTopicSuccess#ShareTopicSuccessSource",
+          "StartCheckInSuccess": "./types/StartCheckInSuccess#StartCheckInSuccessSource",
+          "StartRetrospectiveSuccess": "./types/StartRetrospectiveSuccess#StartRetrospectiveSuccessSource",
           "StartTeamPromptSuccess": "./types/StartTeamPromptSuccess#StartTeamPromptSuccessSource",
           "StripeFailPaymentPayload": "./types/StripeFailPaymentPayload#StripeFailPaymentPayloadSource",
           "Task": "../../database/types/Task#default",
@@ -114,6 +122,7 @@
           "TeamMemberIntegrationAuthOAuth2": "../../postgres/queries/getTeamMemberIntegrationAuth#TeamMemberIntegrationAuth",
           "TeamMemberIntegrations": "./types/TeamMemberIntegrations#TeamMemberIntegrationsSource",
           "TeamPromptMeeting": "../../database/types/MeetingTeamPrompt#default as MeetingTeamPromptDB",
+          "TeamPromptMeetingMember": "../../database/types/TeamPromptMeetingMember#default as TeamPromptMeetingMemberDB",
           "TeamPromptResponse": "../../postgres/queries/getTeamPromptResponsesByIds#TeamPromptResponse",
           "TemplateDimension": "../../database/types/TemplateDimension#default",
           "TimelineEventTeamPromptComplete": "./types/TimelineEventTeamPromptComplete#TimelineEventTeamPromptCompleteSource",
