--- packages/server/graphql/mutations/startRetrospective.ts
@@ -1,151 +1,0 @@
-import {GraphQLID, GraphQLNonNull} from 'graphql'
-import {SubscriptionChannel} from 'parabol-client/types/constEnums'
-import getRethink from '../../database/rethinkDriver'
-import MeetingRetrospective from '../../database/types/MeetingRetrospective'
-import MeetingSettingsRetrospective from '../../database/types/MeetingSettingsRetrospective'
-import RetroMeetingMember from '../../database/types/RetroMeetingMember'
-import generateUID from '../../generateUID'
-import updateMeetingTemplateLastUsedAt from '../../postgres/queries/updateMeetingTemplateLastUsedAt'
-import updateTeamByTeamId from '../../postgres/queries/updateTeamByTeamId'
-import {MeetingTypeEnum} from '../../postgres/types/Meeting'
-import {analytics} from '../../utils/analytics/analytics'
-import {getUserId, isTeamMember} from '../../utils/authorization'
-import publish from '../../utils/publish'
-import standardError from '../../utils/standardError'
-import {GQLContext} from '../graphql'
-import CreateGcalEventInput, {CreateGcalEventInputType} from '../public/types/CreateGcalEventInput'
-import StartRetrospectivePayload from '../types/StartRetrospectivePayload'
-import createGcalEvent from './helpers/createGcalEvent'
-import createNewMeetingPhases from './helpers/createNewMeetingPhases'
-import isStartMeetingLocked from './helpers/isStartMeetingLocked'
-import {IntegrationNotifier} from './helpers/notifications/IntegrationNotifier'
-
-export default {
-  type: new GraphQLNonNull(StartRetrospectivePayload),
-  description: 'Start a new meeting',
-  args: {
-    teamId: {
-      type: new GraphQLNonNull(GraphQLID),
-      description: 'The team starting the meeting'
-    },
-    gcalInput: {
-      type: CreateGcalEventInput,
-      description: 'The gcal event to create. If not provided, no event will be created'
-    }
-  },
-  async resolve(
-    _source: unknown,
-    {teamId, gcalInput}: {teamId: string; gcalInput?: CreateGcalEventInputType},
-    {authToken, socketId: mutatorId, dataLoader}: GQLContext
-  ) {
-    const r = await getRethink()
-    const operationId = dataLoader.share()
-    const subOptions = {mutatorId, operationId}
-    const DUPLICATE_THRESHOLD = 3000
-    // AUTH
-    const viewerId = getUserId(authToken)
-    if (!isTeamMember(authToken, teamId)) {
-      return standardError(new Error('User not on team'), {userId: viewerId})
-    }
-    const unpaidError = await isStartMeetingLocked(teamId, dataLoader)
-    if (unpaidError) return standardError(new Error(unpaidError), {userId: viewerId})
-
-    const meetingType: MeetingTypeEnum = 'retrospective'
-
-    // RESOLUTION
-    const meetingCount = await r
-      .table('NewMeeting')
-      .getAll(teamId, {index: 'teamId'})
-      .filter({meetingType})
-      .count()
-      .default(0)
-      .run()
-
-    const meetingId = generateUID()
-    const phases = await createNewMeetingPhases(
-      viewerId,
-      teamId,
-      meetingId,
-      meetingCount,
-      meetingType,
-      dataLoader
-    )
-    const [team, viewer] = await Promise.all([
-      dataLoader.get('teams').loadNonNull(teamId),
-      dataLoader.get('users').loadNonNull(viewerId)
-    ])
-
-    const organization = await r.table('Organization').get(team.orgId).run()
-    const {showConversionModal} = organization
-
-    const meetingSettings = (await dataLoader
-      .get('meetingSettingsByType')
-      .load({teamId, meetingType})) as MeetingSettingsRetrospective
-    const {
-      id: meetingSettingsId,
-      totalVotes,
-      maxVotesPerGroup,
-      selectedTemplateId,
-      disableAnonymity,
-      videoMeetingURL
-    } = meetingSettings
-    const meeting = new MeetingRetrospective({
-      id: meetingId,
-      teamId,
-      meetingCount,
-      phases,
-      showConversionModal,
-      facilitatorUserId: viewerId,
-      totalVotes,
-      maxVotesPerGroup,
-      disableAnonymity,
-      templateId: selectedTemplateId,
-      videoMeetingURL: videoMeetingURL ?? undefined
-    })
-
-    const template = await dataLoader.get('meetingTemplates').load(selectedTemplateId)
-    await Promise.all([
-      r.table('NewMeeting').insert(meeting).run(),
-      updateMeetingTemplateLastUsedAt(selectedTemplateId, teamId)
-    ])
-
-    // Disallow accidental starts (2 meetings within 2 seconds)
-    const newActiveMeetings = await dataLoader.get('activeMeetingsByTeamId').load(teamId)
-    const otherActiveMeeting = newActiveMeetings.find((activeMeeting) => {
-      const {createdAt, id} = activeMeeting
-      if (id === meetingId || activeMeeting.meetingType !== meetingType) return false
-      return createdAt.getTime() > Date.now() - DUPLICATE_THRESHOLD
-    })
-    if (otherActiveMeeting) {
-      await r.table('NewMeeting').get(meetingId).delete().run()
-      return {error: {message: 'Meeting already started'}}
-    }
-
-    const updates = {
-      lastMeetingType: meetingType
-    }
-    await Promise.all([
-      r
-        .table('MeetingMember')
-        .insert(
-          new RetroMeetingMember({meetingId, userId: viewerId, teamId, votesRemaining: totalVotes})
-        )
-        .run(),
-      updateTeamByTeamId(updates, teamId),
-      videoMeetingURL &&
-        r
-          .table('MeetingSettings')
-          .get(meetingSettingsId)
-          .update({
-            videoMeetingURL: null
-          })
-          .run()
-    ])
-    IntegrationNotifier.startMeeting(dataLoader, meetingId, teamId)
-    analytics.meetingStarted(viewer, meeting, template)
-    const {error} = await createGcalEvent({gcalInput, meetingId, teamId, viewerId, dataLoader})
-    const data = {teamId, meetingId, hasGcalError: !!error?.message}
-    publish(SubscriptionChannel.TEAM, teamId, 'StartRetrospectiveSuccess', data, subOptions)
-    return data
-  }
-}
