+++ packages/server/graphql/public/typeDefs/ActionMeeting.graphql
@@ -1,0 +1,151 @@
+"""
+An action meeting
+"""
+type ActionMeeting implements NewMeeting {
+  """
+  The unique meeting id. shortid.
+  """
+  id: ID!
+
+  """
+  The timestamp the meeting was created
+  """
+  createdAt: DateTime!
+
+  """
+  The id of the user that created the meeting
+  """
+  createdBy: ID!
+
+  """
+  The user that created the meeting
+  """
+  createdByUser: User!
+
+  """
+  The timestamp the meeting officially ended
+  """
+  endedAt: DateTime
+
+  """
+  The location of the facilitator in the meeting
+  """
+  facilitatorStageId: ID!
+
+  """
+  The userId (or anonymousId) of the most recent facilitator
+  """
+  facilitatorUserId: ID!
+
+  """
+  The facilitator team member
+  """
+  facilitator: TeamMember!
+
+  """
+  Is this locked for starter plans?
+  """
+  locked: Boolean!
+
+  """
+  The team members that were active during the time of the meeting
+  """
+  meetingMembers: [ActionMeetingMember!]!
+
+  """
+  The auto-incrementing meeting number for the team
+  """
+  meetingNumber: Int!
+
+  """
+  The id of the meeting series this meeting belongs to
+  """
+  meetingSeriesId: ID
+
+  meetingType: MeetingTypeEnum!
+
+  """
+  The name of the meeting
+  """
+  name: String!
+
+  """
+  The organization this meeting belongs to
+  """
+  organization: Organization!
+
+  """
+  The phases the meeting will go through, including all phase-specific state
+  """
+  phases: [NewMeetingPhase!]!
+
+  """
+  If meeting has a meeting series associated, this is the time the meeting will end
+  """
+  scheduledEndTime: DateTime
+
+  """
+  true if should show the org the conversion modal, else false
+  """
+  showConversionModal: Boolean!
+
+  """
+  The OpenAI generated summary of all the content in the meeting, such as reflections, tasks, and comments. Undefined if the user doesnt have access to the feature or it's unavailable in this meeting type`
+  """
+  summary: String
+
+  """
+  The time the meeting summary was emailed to the team
+  """
+  summarySentAt: DateTime
+
+  """
+  foreign key for team
+  """
+  teamId: ID!
+
+  """
+  The team that ran the meeting
+  """
+  team: Team!
+
+  """
+  The last time a meeting was updated (stage completed, finished, etc)
+  """
+  updatedAt: DateTime
+
+  """
+  The action meeting member of the viewer
+  """
+  viewerMeetingMember: ActionMeetingMember
+
+  """
+  A single agenda item
+  """
+  agendaItem(agendaItemId: ID!): AgendaItem
+
+  """
+  The number of agenda items generated in the meeting
+  """
+  agendaItemCount: Int!
+
+  """
+  All of the agenda items for the meeting
+  """
+  agendaItems: [AgendaItem!]!
+
+  """
+  The number of comments generated in the meeting
+  """
+  commentCount: Int!
+
+  """
+  The number of tasks generated in the meeting
+  """
+  taskCount: Int!
+
+  """
+  The tasks created within the meeting
+  """
+  tasks: [Task!]!
+}
