+++ packages/server/graphql/public/typeDefs/startCheckIn.graphql
@@ -1,0 +1,31 @@
+"""
+Return object for StartCheckInPayload
+"""
+union StartCheckInPayload = StartCheckInSuccess | ErrorPayload
+
+type StartCheckInSuccess {
+  meeting: ActionMeeting!
+  meetingId: ID!
+  team: Team!
+  hasGcalError: Boolean!
+}
+
+extend type Mutation {
+  """
+  Start a new meeting
+  """
+  startCheckIn(
+    """
+    The team starting the meeting
+    """
+    teamId: ID
+    """
+    The gcal input if creating a gcal event
+    """
+    gcalInput: CreateGcalEventInput
+    """
+    One-on-one ad-hoc team input
+    """
+    oneOnOneTeamInput: CreateOneOnOneTeamInput
+  ): StartCheckInPayload!
+}
