+++ packages/server/graphql/public/types/StartRetrospectiveSuccess.ts
@@ -1,0 +1,18 @@
+import MeetingRetrospective from '../../../database/types/MeetingRetrospective'
+import {StartRetrospectiveSuccessResolvers} from '../resolverTypes'
+
+export type StartRetrospectiveSuccessSource = {
+  meetingId: string
+  teamId: string
+}
+
+const StartRetrospectiveSuccess: StartRetrospectiveSuccessResolvers = {
+  meeting: ({meetingId}, _args: unknown, {dataLoader}) => {
+    return dataLoader.get('newMeetings').load(meetingId) as Promise<MeetingRetrospective>
+  },
+  team: ({teamId}, _args: unknown, {dataLoader}) => {
+    return dataLoader.get('teams').loadNonNull(teamId)
+  }
+}
+
+export default StartRetrospectiveSuccess
