+++ packages/server/graphql/rootTypes.ts
@@ -1,4 +1,3 @@
-import ActionMeeting from './types/ActionMeeting'
 import ActionMeetingMember from './types/ActionMeetingMember'
 import ActionMeetingSettings from './types/ActionMeetingSettings'
 import AgendaItemsPhase from './types/AgendaItemsPhase'
@@ -17,7 +16,6 @@ import JiraDimensionField from './types/JiraDimensionField'
 import PokerMeetingSettings from './types/PokerMeetingSettings'
 import ReflectPhase from './types/ReflectPhase'
 import RenamePokerTemplatePayload from './types/RenamePokerTemplatePayload'
-import RetrospectiveMeeting from './types/RetrospectiveMeeting'
 import RetrospectiveMeetingMember from './types/RetrospectiveMeetingMember'
 import RetrospectiveMeetingSettings from './types/RetrospectiveMeetingSettings'
 import SetMeetingSettingsPayload from './types/SetMeetingSettingsPayload'
@@ -26,7 +24,6 @@ import SuggestedActionInviteYourTeam from './types/SuggestedActionInviteYourTeam
 import SuggestedActionTryActionMeeting from './types/SuggestedActionTryActionMeeting'
 import SuggestedActionTryRetroMeeting from './types/SuggestedActionTryRetroMeeting'
 import SuggestedActionTryTheDemo from './types/SuggestedActionTryTheDemo'
-import TeamPromptMeeting from './types/TeamPromptMeeting'
 import TeamPromptMeetingMember from './types/TeamPromptMeetingMember'
 import TeamPromptMeetingSettings from './types/TeamPromptMeetingSettings'
 import TeamPromptResponsesPhase from './types/TeamPromptResponsesPhase'
@@ -53,10 +50,8 @@ const rootTypes = [
   TeamPromptResponsesPhase,
   GenericMeetingPhase,
   EstimatePhase,
-  ActionMeeting,
   ActionMeetingMember,
   PokerMeetingSettings,
-  RetrospectiveMeeting,
   RetrospectiveMeetingMember,
   RetrospectiveMeetingSettings,
   SetMeetingSettingsPayload,
@@ -71,7 +66,6 @@ const rootTypes = [
   TimelineEventCompletedActionMeeting,
   TimelineEventPokerComplete,
   ActionMeetingSettings,
-  TeamPromptMeeting,
   TeamPromptMeetingMember,
   TeamPromptMeetingSettings,
   Comment,
